Allow users to optionally let Xwayland to eavesdrop on their keyboard
It's rather popular for voice communication applications to support
Push-to-Talk. This means that the process itself expects to get all of
the system input. This behaviour albeit sound does not work on Wayland
systems.

This commit adds an option to let legacy X11 applications that assume
they will be getting all information to do so until we are ready to
receive these use-cases on Wayland.