kcm/kwinrules: Improve button behavior when adding properties
- Make the button on each delegate visible not only on hover
to be more consistent with current status everywhere else.

- Change the icon when the property is already added and this
will just apply the suggestion. Using `edit-redo` as a lack of
a better `edit-apply` or `edit-replace` icon.

- Keep properties sheet open when clicking on the button to add
a new property. This allows to add several properties in a row.
Clicking on the full row will keep the previous behavior for the
simple case, closing the sheet so the new property can be edited
right away

Previously it worked the same for both items, so the button was
merely informative and the sheet would remain open only after
detecting some window properties, which was a very hidden and
confusing pattern.

Not CCBUGing on purpose, since while the idea is good, it comes
from a rant-formed catch-it-all bug report.