backends/drm: offset rotation with panel orientation property
The panel orientation property tells compositors about the orientation that
displays are built into specific devices.
In order to support these devices with rotated displays best, offset the
rotation the user can set by the rotation set by the property. With that,
normal rotation set in display settings will be what the user assumes to
be normal.