effects/slide: Refactor current grid position handling
At the moment, the slide effect maintains m_currentPosition with the
current position even if it's inactive. This is slightly error-prone as
it is easy to overlook a case when m_currentPosition has to be updated.
In fact, the slide effect doesn't handle the case when the number of
rows changes.

It's less error prone if the desktop grid position is calculated on
demand.